+++ boilerplates/nuxt/components/SfButton/SfButtonPrimary.vue
@@ -1,33 +1,24 @@
 <template>
   <button
-    v-on-key-stroke="keyHandler"
-    ref="el"
+    v-focus
+    :disabled="disabled"
     class="inline-flex items-center justify-center cursor-pointer bg-green-500 hover:bg-green-600 active:bg-green-700 disabled:bg-gray-200 text-white disabled:text-gray-500 disabled:opacity-50 uppercase font-body text-base font-medium px-[16px] py-2 rounded-md shadow hover:shadow-md disabled:shadow-none disabled:cursor-not-allowed outline-violet"
-    :class="{'outline-none': styleFocus}"
   >
     <slot />
   </button>
 </template>
 
 <script>
-import { vOnKeyStroke } from '@vueuse/components'
-
+import { focus } from '../../../utils/focus-directive.js'
 export default {
-  name: 'SfButtonPrimary',
-  data() {
-    return {
-      styleFocus: true
-    }
-  },
+  name: 'ButtonPrimary',
   directives: {
-    vOnKeyStroke
+    focus
   },
-  methods: {
-    keyHandler(e, el) {
-      if (e === 'Tab') {
-        this.styleFocus = false
-        el.focus()
-      }
+  props: {
+    disabled: {
+      type: Boolean,
+      default: false
     }
   }
 }