+++ shared/docs/v0.0.1/changelog.md
@@ -1,6 +1,7 @@
 # v0.0.1
 
 ## 🚀 Features:
+- add VueUse package
 
 ## 🐛 Fixes: